GCM work in progress.
It looks like something weird is going on the decryption side, even
before we're able to pass through the tag. The DecryptUpdate operation
is reporting an error but isn't throwing anything into the error stack
for us to look at. Stashing on a branch so I can continue with some code
cleanup in the older CBC based mode.